Separate vector utils from analysis
